gcp/cloud_run_functions/run_intake_confirmation/requirements.txt
aiohappyeyeballs==2.6.1 ; python_version == "3.12"
aiohttp-retry==2.9.1 ; python_version == "3.12"
aiohttp==3.13.3 ; python_version == "3.12"
aiosignal==1.4.0 ; python_version == "3.12"
annotated-types==0.7.0 ; python_version == "3.12"
anyio==4.12.1 ; python_version == "3.12"
attrs==25.4.0 ; python_version == "3.12"
blinker==1.9.0 ; python_version == "3.12"
certifi==2026.1.4 ; python_version == "3.12"
cffi==2.0.0 ; python_version == "3.12" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.4 ; python_version == "3.12"
click==8.1.8 ; python_version == "3.12"
cloudevents==1.12.0 ; python_version == "3.12"
colorama==0.4.6 ; python_version == "3.12" and platform_system == "Windows"
cryptography==46.0.5 ; python_version == "3.12"
deprecation==2.1.0 ; python_version == "3.12"
flask==3.1.2 ; python_version == "3.12"
frozenlist==1.8.0 ; python_version == "3.12"
functions-framework==3.10.1 ; python_version == "3.12"
google-api-core==2.29.0 ; python_version == "3.12"
google-auth==2.48.0 ; python_version == "3.12"
google-cloud-bigquery==3.39.0 ; python_version == "3.12"
google-cloud-core==2.5.0 ; python_version == "3.12"
google-cloud-pubsub==2.35.0 ; python_version == "3.12"
google-crc32c==1.8.0 ; python_version == "3.12"
google-resumable-media==2.8.0 ; python_version == "3.12"
googleapis-common-protos==1.72.0 ; python_version == "3.12"
grpc-google-iam-v1==0.14.3 ; python_version == "3.12"
grpcio-status==1.78.0 ; python_version == "3.12"
grpcio==1.78.0 ; python_version == "3.12"
gunicorn==25.1.0 ; python_version == "3.12"
h11==0.16.0 ; python_version == "3.12"
idna==3.11 ; python_version == "3.12"
importlib-metadata==8.7.1 ; python_version == "3.12"
itsdangerous==2.2.0 ; python_version == "3.12"
jinja2==3.1.6 ; python_version == "3.12"
markupsafe==3.0.3 ; python_version == "3.12"
multidict==6.7.1 ; python_version == "3.12"
opentelemetry-api==1.39.1 ; python_version == "3.12"
opentelemetry-sdk==1.39.1 ; python_version == "3.12"
opentelemetry-semantic-conventions==0.60b1 ; python_version == "3.12"
packaging==26.0 ; python_version == "3.12"
propcache==0.4.1 ; python_version == "3.12"
proto-plus==1.27.1 ; python_version == "3.12"
protobuf==6.33.5 ; python_version == "3.12"
pyasn1-modules==0.4.2 ; python_version == "3.12"
pyasn1==0.6.2 ; python_version == "3.12"
pycparser==3.0 ; platform_python_implementation != "PyPy" and implementation_name != "PyPy" and python_version == "3.12"
pydantic-core==2.41.5 ; python_version == "3.12"
pydantic-settings==2.13.0 ; python_version == "3.12"
pydantic==2.12.5 ; python_version == "3.12"
pyjwt==2.11.0 ; python_version == "3.12"
python-dateutil==2.9.0.post0 ; python_version == "3.12"
python-dotenv==1.2.1 ; python_version == "3.12"
pyyaml==6.0.3 ; python_version == "3.12"
requests==2.32.5 ; python_version == "3.12"
rsa==4.9.1 ; python_version == "3.12"
six==1.17.0 ; python_version == "3.12"
starlette==0.52.1 ; python_version == "3.12"
twilio==9.10.1 ; python_version == "3.12"
typing-extensions==4.15.0 ; python_version == "3.12"
typing-inspection==0.4.2 ; python_version == "3.12"
urllib3==2.6.3 ; python_version == "3.12"
uvicorn-worker==0.4.0 ; python_version == "3.12"
uvicorn==0.41.0 ; python_version == "3.12"
watchdog==6.0.0 ; python_version == "3.12"
werkzeug==3.1.5 ; python_version == "3.12"
yarl==1.22.0 ; python_version == "3.12"
zipp==3.23.0 ; python_version == "3.12"
